requirements.txt
torch>=2.0.0  
ultralytics
flask
pillow
numpy
pandas
seaborn
matplotlib
ipywidgets  